Let PHP do the work of decoding entities
Rather than maintaining lists of entities to search for and replace, let
PHP do the work, and just maintain a list of exceptions where necessary.

Exceptions are for where the code previously supported windows-cp1252
entities and for special support for ampersands.

This also means that the proper character will be used, for example
rather than &copy; being decoded into (c) it will be decoded into ©.