Increment diff counter in blockqute parsing (Fixes #37)
The blockquote counter should be incremented when parsing quoted text,
rather than set to the length of the quote. If there are multiple
blockquotes in the text then the position counter must be tracked
additively and cannot be assumed to start from zero each time.

Thanks to @voku, who provided this patch. I've just grabbed his patch, and
removed his local customisations for UTF8 versions of strlen.